Move analysis classes into own package
